moved autonomous code under frc folder
