+++ src/main/java/frc/autonomous/Command.java
@@ -1,4 +1,4 @@
-package frc.robot.autonomous;
+package frc.autonomous;
 
 public interface Command {
