+++ src/main/java/frc/autonomous/Mission.java
@@ -1,4 +1,4 @@
-package frc.robot.autonomous;
+package frc.autonomous;
 
 import java.util.ArrayList;
 import java.util.Arrays;
